requirements.txt
numpy==1.26.4
ultralytics
opencv-python-headless
watchdog
pyyaml
onnxruntime-gpu==1.17.0
